moves headers of the public interface into an sg14 subdirectory
- avoids confusion between:
  - standard header, type_traits
  - and wannabe project header, type_traits.h
