+++ include/sg14/bits/common.h
@@ -1,0 +1,24 @@
+//          Copyright John McFarlane 2015 - 2016.
+// Distributed under the Boost Software License, Version 1.0.
+//    (See accompanying file ../LICENSE_1_0.txt or copy at
+//          http://www.boost.org/LICENSE_1_0.txt)
+
+// definitions that are directly required by more than one header of the API
+
+#if !defined(_SG14_COMMON_H)
+#define _SG14_COMMON_H 1
+
+namespace sg14 {
+    namespace _impl {
+        ////////////////////////////////////////////////////////////////////////////////
+        // sg14::_impl::max
+
+        template<class T>
+        constexpr T max(T a, T b)
+        {
+            return (a<b) ? b : a;
+        }
+    }
+}
+
+#endif  // _SG14_COMMON_H
