+++ src/common/sample_functions.h
@@ -4,7 +4,7 @@
 //  (See accompanying file ../../LICENSE_1_0.txt or copy at
 //          http://www.boost.org/LICENSE_1_0.txt)
 
-#include <fixed_point.h>
+#include <sg14/fixed_point.h>
 
 #include <cmath>
