+++ src/test/fixed_point_built_in.cpp
@@ -4,9 +4,9 @@
 //  (See accompanying file ../../LICENSE_1_0.txt or copy at
 //          http://www.boost.org/LICENSE_1_0.txt)
 
-#include <cinttypes>
+#include <sg14/fixed_point.h>
 
-#include <fixed_point.h>
+#include <cinttypes>
 
 #define TEST_NATIVE_OVERFLOW
 #define TEST_LABEL built_in_
