+++ src/test/fixed_point_native_integer.cpp
@@ -5,7 +5,7 @@
 //          http://www.boost.org/LICENSE_1_0.txt)
 
 #include "integer.h"
-#include <fixed_point.h>
+#include <sg14/fixed_point.h>
 
 #define TEST_NATIVE_OVERFLOW
 #define TEST_LABEL native_integer_
