+++ src/test/snippets.cpp
@@ -4,8 +4,8 @@
 //  (See accompanying file ../../LICENSE_1_0.txt or copy at
 //          http://www.boost.org/LICENSE_1_0.txt)
 
-#include <fixed_point.h>
-#include <elastic.h>
+#include <sg14/fixed_point.h>
+#include <sg14/elastic.h>
 
 using namespace sg14;
