+++ src/test/type_traits.cpp
@@ -4,7 +4,7 @@
 //  (See accompanying file ../../LICENSE_1_0.txt or copy at
 //          http://www.boost.org/LICENSE_1_0.txt)
 
-#include <type_traits.h>
+#include <sg14/type_traits.h>
 
 using sg14::_type_traits_impl::first_fit_t;
 using std::is_same;
